requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
matplotlib
opencv-python
gradio
transformers
